Refined edit, now auto updates upon save instead of refreshing page
