+++ src/scripts/events.js
@@ -1,9 +1,9 @@
-//--Radio Filter --//
+// //--Radio Filter --//
 
-radioBtns = document.getElementsByName("moodFilter");
+// radioBtns = document.getElementsByName("moodFilter");
 
-radioBtns.ForEach(radioBtn => {
-    radioBtns.addEventListener("click", () => {
-        const mood = event.target.value
-    })
-});
+// radioBtns.ForEach(radioBtn => {
+//     radioBtns.addEventListener("click", () => {
+//         const mood = event.target.value
+//     })
+// });
